Separate personal information fields into different component
